Support multiple/nested sheets and sheet instances
