+++ packages/r3f/src/SheetProvider.tsx
@@ -1,16 +1,16 @@
 import React, {
   createContext,
   useContext,
+  useEffect,
   useLayoutEffect,
-  useState,
 } from 'react'
 import {useThree} from '@react-three/fiber'
 import type {ISheet} from '@theatre/core'
 import {bindToCanvas} from './store'
 
-const ctx = createContext<{sheet: ISheet | undefined} | undefined>(undefined)
+const ctx = createContext<{sheet: ISheet}>(undefined!)
 
-const useWrapperContext = (): {sheet: ISheet | undefined} => {
+const useWrapperContext = (): {sheet: ISheet} => {
   const val = useContext(ctx)
   if (!val) {
     throw new Error(
@@ -25,23 +25,21 @@ export const useCurrentSheet = (): ISheet | undefined => {
 }
 
 const SheetProvider: React.FC<{
-  getSheet: () => ISheet
-}> = (props) => {
+  sheet: ISheet
+}> = ({sheet, children}) => {
   const {scene, gl} = useThree((s) => ({scene: s.scene, gl: s.gl}))
-  const [sheet, setSheet] = useState<ISheet | undefined>(undefined)
+
+  useEffect(() => {
+    if (!sheet || sheet.type !== 'Theatre_Sheet_PublicAPI') {
+      throw new Error(`sheet in <Wrapper sheet={sheet}> has an invalid value`)
+    }
+  }, [sheet])
 
   useLayoutEffect(() => {
-    const sheet = props.getSheet()
-    if (!sheet || sheet.type !== 'Theatre_Sheet_PublicAPI') {
-      throw new Error(
-        `getSheet() in <Wrapper getSheet={getSheet}> has returned an invalid value`,
-      )
-    }
-    setSheet(sheet)
     bindToCanvas({gl, scene})
   }, [scene, gl])
 
-  return <ctx.Provider value={{sheet}}>{props.children}</ctx.Provider>
+  return <ctx.Provider value={{sheet}}>{children}</ctx.Provider>
 }
 
 export default SheetProvider
